auth: use unreadable-fancy-python-that-makes-you-look-smart to make coverage happy
instead of having a simple for loop to find the highest user id, I use
this gross thing with max which makes sure you have to scratch your head
to understand it. I even had to write a comment for it... Yikes. This
will be fixed by #70

We got 100% coverage on auth now though (the problem was that ids were
sorted, and so the if n > max condition was always true, making coverage
unhappy. And we've got no way of removing users, and it was too painful
to use random.shuffle just to please coverage)